bugfix:fix incremental build error

+++ matrix/matrix-android/matrix-gradle-plugin/src/main/java/com/tencent/matrix/trace/transform/MatrixTraceTransform.java
@@ -414,7 +414,7 @@ public class MatrixTraceTransform extends Transform {
 
             final File jarInput = inputJar.getFile();
             final File jarOutput = new File(traceClassOut, getUniqueJarName(jarInput));
-            if (jarOutput.exists()) {
+            if (!jarInput.exists() && jarOutput.exists()) {
                 jarOutput.delete();
             }
             if (!jarOutput.getParentFile().exists()) {
